created constants and reducer files
